requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
bcrypt
python-jose[cryptography]
python-multipart
python-dotenv